feat: 2u-143 refactor messages
