+++ src/course-updates/course-handouts/messages.js
@@ -1,0 +1,14 @@
+import { defineMessages } from '@edx/frontend-platform/i18n';
+
+const messages = defineMessages({
+  handoutsTitle: {
+    id: 'course-authoring.course-updates.handouts.title',
+    defaultMessage: 'Course Handouts',
+  },
+  editButton: {
+    id: 'course-authoring.course-updates.actions.edit',
+    defaultMessage: 'Edit',
+  },
+});
+
+export default messages;
