+++ src/course-updates/messages.js
@@ -13,77 +13,9 @@ const messages = defineMessages({
     id: 'course-authoring.course-updates.section-info',
     defaultMessage: 'Use course updates to notify students of important dates or exams, highlight particular discussions in the forums, announce schedule changes, and respond to student questions. You add or edit updates in HTML.',
   },
-  handoutsTitle: {
-    id: 'course-authoring.course-updates.handouts.title',
-    defaultMessage: 'Course Handouts',
-  },
-  editButton: {
-    id: 'course-authoring.course-updates.actions.edit',
-    defaultMessage: 'Edit',
-  },
-  deleteButton: {
-    id: 'course-authoring.course-updates.actions.delete',
-    defaultMessage: 'Delete',
-  },
-  saveButton: {
-    id: 'course-authoring.course-updates.actions.save',
-    defaultMessage: 'Save',
-  },
-  postButton: {
-    id: 'course-authoring.course-updates.actions.post',
-    defaultMessage: 'Post',
-  },
-  cancelButton: {
-    id: 'course-authoring.course-updates.actions.cancel',
-    defaultMessage: 'Cancel',
-  },
   newUpdateButton: {
     id: 'course-authoring.course-updates.actions.new-update',
     defaultMessage: 'New Update',
-  },
-  okButton: {
-    id: 'course-authoring.course-updates.button.ok',
-    defaultMessage: 'Ok',
-  },
-  dateInvalidText: {
-    id: 'course-authoring.course-updates.date-invalid',
-    defineMessage: 'Action required: Enter a valid date.',
-  },
-  deleteModalTitle: {
-    id: 'course-authoring.course-updates.delete-modal.title',
-    defaultMessage: 'Are you sure you want to delete this update?',
-  },
-  deleteModalDescription: {
-    id: 'course-authoring.course-updates.delete-modal.description',
-    defaultMessage: 'This action cannot be undone.',
-  },
-  updateModalDate: {
-    id: 'course-authoring.course-updates.update-modal.date',
-    defaultMessage: 'Date:',
-  },
-  updateModalInValid: {
-    id: 'course-authoring.course-updates.update-modal.date',
-    defaultMessage: 'Date:',
-  },
-  updateModalCalendarAltText: {
-    id: 'course-authoring.course-updates.update-modal.calendar-alt-text',
-    defaultMessage: 'Calendar for datepicker input',
-  },
-  updateModalErrorAltText: {
-    id: 'course-authoring.course-updates.update-modal.error-alt-text',
-    defaultMessage: 'Error icon',
-  },
-  addNewUpdateTitle: {
-    id: 'course-authoring.course-updates.update-modal.new-update-title',
-    defaultMessage: 'Add new update',
-  },
-  editUpdateTitle: {
-    id: 'course-authoring.course-updates.update-modal.edit-update-title',
-    defaultMessage: 'Edit updates',
-  },
-  editHandoutsTitle: {
-    id: 'course-authoring.course-updates.update-modal.edit-handouts-title',
-    defaultMessage: 'Edit handouts',
   },
 });
