+++ src/course-updates/update-modal/messages.js
@@ -1,0 +1,46 @@
+import { defineMessages } from '@edx/frontend-platform/i18n';
+
+const messages = defineMessages({
+  updateModalDate: {
+    id: 'course-authoring.course-updates.update-modal.date',
+    defaultMessage: 'Date:',
+  },
+  updateModalInValid: {
+    id: 'course-authoring.course-updates.update-modal.date',
+    defaultMessage: 'Date:',
+  },
+  updateModalCalendarAltText: {
+    id: 'course-authoring.course-updates.update-modal.calendar-alt-text',
+    defaultMessage: 'Calendar for datepicker input',
+  },
+  updateModalErrorAltText: {
+    id: 'course-authoring.course-updates.update-modal.error-alt-text',
+    defaultMessage: 'Error icon',
+  },
+  addNewUpdateTitle: {
+    id: 'course-authoring.course-updates.update-modal.new-update-title',
+    defaultMessage: 'Add new update',
+  },
+  editUpdateTitle: {
+    id: 'course-authoring.course-updates.update-modal.edit-update-title',
+    defaultMessage: 'Edit updates',
+  },
+  editHandoutsTitle: {
+    id: 'course-authoring.course-updates.update-modal.edit-handouts-title',
+    defaultMessage: 'Edit handouts',
+  },
+  saveButton: {
+    id: 'course-authoring.course-updates.actions.save',
+    defaultMessage: 'Save',
+  },
+  postButton: {
+    id: 'course-authoring.course-updates.actions.post',
+    defaultMessage: 'Post',
+  },
+  cancelButton: {
+    id: 'course-authoring.course-updates.actions.cancel',
+    defaultMessage: 'Cancel',
+  },
+});
+
+export default messages;
